requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
-e git+ssh://git@github.com/michaeleliot/Discovery-Engine.git@b621c87597a524b6a9de8ae7781cc80bc9cc4a81#egg=discovery_engine
executing==2.2.0
frozenlist==1.7.0
genai==2.1.0
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.6.4
numpy==2.3.2
openai==0.27.10
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
